refactor: split main.go monolith into 16 focused files
Break cmd/mnemonic/main.go (3407 lines) into logical files:
- main.go (241 lines): router, constants, printUsage
- serve.go (730): daemon serve command
- dedup_cli.go (368): dedup and reset-patterns commands
- cycle.go (263): meta-cycle, dream-cycle, mcp, autopilot
- status.go (251): status command and helpers
- runtime.go (233): config builders, LLM provider factory
- daemon.go (207): start/stop/restart commands
- memory_cli.go (182): remember, recall, consolidate
- purge.go (175): purge and cleanup commands
- diagnose.go (174): diagnose command
- export.go (143): export, import, backup
- watch.go (134): watch command
- web.go (124): agent web server
- update.go (90): check-update, update, generate-token
- insights.go (89): insights command
- restore.go (73): restore command
- install.go (60): install/uninstall commands
- util.go (9): truncate helper

All files remain in package main. No function signatures changed.

Addresses finding #10 from yield audit (issue #355).
