+++ cmd/mnemonic/util.go
@@ -1,0 +1,9 @@
+package main
+
+// truncate shortens a string to maxLen with ellipsis.
+func truncate(s string, maxLen int) string {
+	if len(s) <= maxLen {
+		return s
+	}
+	return s[:maxLen-3] + "..."
+}
